Fliter/map type aliases before code generation
Ocamlnet now has a fast constructor for string arrays.  Orpc already generates correct code for the new constructor when the type is a string. This commit will expand orpc to cover types that are aliases for strings (previously, that would cause a run time error in the client).

Here's an example that was resulting in a runtime error, but works with this commit:

    type key = string
    type value = string

    module type Abstract =
    sig
      type _r

      val get_value_list : key -> value list
      val get_value_array : key -> value array
    end

This commit gets the desired result by mapping type aliases for primitives (and aliases for aliases of primitives) to their actual primitive so that the code generation phase can properly detect the primitive type.